logicaffeine_web

Module unlock

Module Unlock Logic

Rules:

First module in each era is always unlocked

Subsequent modules unlock when the previous module is completed

Last two modules in each era are locked until at least one module has 100% completion

Enums

ModuleState

State of a module for the user

Functions

check_module_unlocked

Check if a specific module is unlocked for the user

get_locked_module_ids

Get list of locked module IDs for an era

get_module_state

Get the current state of a module for the user

has_perfect_module

Check if any module in the era has 100% completion